Match signature length handling

+++ crypto/hblk_crypto.h
@@ -21,7 +21,7 @@
 typedef struct sig_s
 {
 	uint8_t sig[SIG_MAX_LEN];
-	size_t len;
+	uint8_t len;
 } sig_t;
 
 uint8_t *sha256(int8_t const *s, size_t len,
